2023, Day 3: added part 2 solution
This is pretty brute force. We materialize the token stream, split off
the gear symbols and use them as keys for a hashmap of vectors. We then
iterate the rest of the symbols and push part numbers onto those vectors
if the part number is adjacent. From there, I just map and fold to get
the final answer.