2023, Day 4: part1 & part2 solutions
Part 1 is a little more explicit in checking input validity, but the
approach for part 1 and 2 are basically the same. Discard everything
from the colon forward on each line, then split on the vertical bar and
parse the resulting strings into a hash set and vector, respectively.
The hash set contains the winners and the vector has the list of cards
we have.

Part 2 parses the data the same way, but it does it in one big jackass
map statement for maximum unreadability.

Once the vector of (hashmap, vector) pairs is established, both parts
are pretty simple.