2023, Day 5: part2 solution
First observe that part2 is just a generalization of part1. The basic
rules are the same, it's just how we generate seeds that's a bit
different. I probably should have made this two patches, but I started
by rewriting part1 as the special case where the seed 'run length' is
implicitly one and then debugged that.

After that, I factored out the seed parse code into a separate function,
which gave me a chance to experiment with functions as arguments. Then,
part2 is just a different seed parse function.