2023, Day 6 solution
The majority of the work here was to going through the contortions to
get Rust to parse the input, especially in part 2. Having the data input
be two semantically different lines instead of each line corresponding
to a different race definitely makes the parsing more complicated.

For part 1, I initially did a brute force solution, but then part 2 made
that non-practical, so I went back and refreshed my algebra and how to
solve quadratic equations. Thanks Ms. Jennings! With the quadratic
formula implemented, I debugged on the part 1 input (the equality on the
third race threw me for a hot minute) so part 2 just worked the first time.