2023: Day 8
First part is straightforward. Parse the data and walk the graph until
you hit the end node. I did this a different way than seen here
originally, but refactored once the second part made things more
complicated.

Second part is potentially very complicated but, by happenstance or
benevolence, the distance of the terminal node from the start is always
the same as the cycle length for each starting node. That allows us to
use least common multiple to find when the paths will all converge on
terminal nodes together. Otherwise the math could get a bit hairy. I
flailed on that for a bit before trying the simple solution.